requirements.txt
grpcio
grpcio-reflection
numpy
Pillow
torch
torchvision